Add files via upload
What we have so far, what we have left is the interpolation (handling the NA values). I just replaced them for the time being. Also, on the overview section, have different colors for the three continents, and if you hover over a point, it should display the car name. I know how to implement the hover stuff, but for some reason the graph library doesn't let it show.